+++ jautograd-compute/src/main/java/com/ashutoshwad/utils/jautograd/MaxForwardOperation.java
@@ -21,6 +21,6 @@ class MaxForwardOperation extends ForwardComputeOperation {
                 }
             }
         }
-        result.setValue(0, 0, max);
+        result.setValue(0, 0, (float)max);
     }
 }
